+++ lib/assent/strategies/oidc.ex
@@ -237,7 +237,7 @@ defmodule Assent.Strategy.OIDC do
   def validate_id_token(config, id_token) do
     expected_alg = Config.get(config, :id_token_signed_response_alg, "RS256")
 
-    with {:ok, openid_config} <- Config.fetch(config, :openid_configuration),
+    with {:ok, openid_config} <- openid_configuration(config),
          {:ok, client_id}     <- Config.fetch(config, :client_id),
          {:ok, issuer}        <- fetch_from_openid_config(openid_config, "issuer"),
          {:ok, jwt}           <- verify_jwt(id_token, openid_config, config),
@@ -370,7 +370,7 @@ defmodule Assent.Strategy.OIDC do
 
   @spec fetch_userinfo(Config.t(), map()) :: {:ok, map()} | {:error, term()}
   def fetch_userinfo(config, token) do
-    with {:ok, openid_config} <- Config.fetch(config, :openid_configuration),
+    with {:ok, openid_config} <- openid_configuration(config),
          {:ok, userinfo_url}  <- fetch_from_openid_config(openid_config, "userinfo_endpoint"),
          {:ok, claims}        <- fetch_from_userinfo_endpoint(config, openid_config, token, userinfo_url),
          :ok                  <- validate_userinfo_sub(config, token["id_token"], claims) do